Move PoseInfo to its own file.
