+++ Pose.cs
@@ -16,20 +16,6 @@ namespace FbxSharp
         readonly ChangeNotifyList<PoseInfo> poseInfos = new ChangeNotifyList<PoseInfo>();
         public readonly CollectionView<PoseInfo, PoseInfo> PoseInfos;
 
-        public struct PoseInfo
-        {
-            public PoseInfo(FbxNode node, FbxMatrix matrix, bool matrixIsLocal=false)
-            {
-                Node = node;
-                Matrix = matrix;
-                MatrixIsLocal = matrixIsLocal;
-            }
-
-            public readonly FbxNode Node;
-            public readonly FbxMatrix Matrix;
-            public readonly bool MatrixIsLocal;
-        }
-
         public void SetIsBindPose(bool pIsBindPose)
         {
             isBindPose = pIsBindPose;
